+++ packages/dialect-react-ui/components/common/primitives.tsx
@@ -188,12 +188,11 @@ export function Accordion(props: {
   return (
     <div className={props?.className}>
       <div
-        // onClick={() => setExpanded((prev) => !prev)}
         className={cs(
           textStyles.bigText,
           'dt-w-full dt-flex dt-justify-between dt-mb-1'
         )}
-      >
+      > 
         {props.title}
         {/* <IconButton
           icon={<icons.chevron />}
@@ -209,6 +208,7 @@ export function Accordion(props: {
     </div>
   );
 }
+
 
 export const useBalance = () => {
   const { wallet, program } = useApi();